+++ sound/soc/mediatek/common/auddrv_underflow_mach.c
@@ -1,17 +1,19 @@
 /*
- * Copyright (C) 2007 The Android Open Source Project
+ * Copyright (C) 2015 MediaTek Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 as
+ * published by the Free Software Foundation.
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * You should have received a copy of the GNU General Public License
+ * along with this program
+ * If not, see <http://www.gnu.org/licenses/>.
  */
 /*******************************************************************************
  *
@@ -173,7 +175,7 @@ void Auddrv_CheckInterruptTiming(void)
 			     __func__, Irq_time_t2, Irq_time_t1, Irq_time_t1 - Irq_time_t2,
 			     mDL1_Interrupt_Interval_Limit);
 			Irq_time_t2 = Irq_time_t1 - Irq_time_t2;
-			if (Irq_time_t2 > mDL1_Interrupt_Interval_Limit * 1000000) {
+			if (Irq_time_t2 > (unsigned long long)mDL1_Interrupt_Interval_Limit * 1000000) {
 				pr_debug
 				    ("%s interrupt may be blocked Irq_time_t2 = %llu Interval_Limit = %d\n",
 				     __func__, Irq_time_t2, mDL1_Interrupt_Interval_Limit);
